Changed SubResourceManager for ResourceManager. Fixed warnings.

+++ ByteEngine/src/ByteEngine/Application/StackAllocator.h
@@ -119,7 +119,7 @@ protected:
 	};
 
 	const uint64 blockSize{ 0 };
-	GTSL::Atomic<uint64> stackIndex{ 0 };
+	GTSL::Atomic<uint32> stackIndex{ 0 };
 	GTSL::Vector<GTSL::Vector<Block>> stacks;
 	GTSL::Vector<GTSL::Mutex> stacksMutexes;
 	GTSL::AllocatorReference* allocatorReference{ nullptr };
